un-comment and test conditional version of deleteRepair ctrl- SUCCESS!!!

+++ controllers/repairs.js
@@ -57,22 +57,22 @@ export async function update(req, res) {
 
 async function deleteRepair(req, res) {
   try {
-    const repair = await Repair.findByIdAndDelete(req.params.repairId)
-    const profile = await Profile.findById(req.user.profile)
-    profile.repairs.remove({ _id: req.params.repairId })
-    await profile.save()
-    res.json(repair)
+    // const repair = await Repair.findByIdAndDelete(req.params.repairId)
+    // const profile = await Profile.findById(req.user.profile)
+    // profile.repairs.remove({ _id: req.params.repairId })
+    // await profile.save()
+    // res.json(repair)
     // --- below code throws error - same as above but wanted to add barrier to keep anyone but the owner of the repair from being able to delete it ---
-    // const repair = await Repair.findById(req.params.repairId)
-    // if (repair.owner.equals(req.user.profile)) {
-    //   await Repair.findByIdAndDelete(req.params.repairId)
-    //   const profile = await Profile.findById(req.user.profile)
-    //   profile.repairs.remove({ _id: req.params.repairId })
-    //   await profile.save()
-    //   res.json(repair)
-    // } else {
-    //   throw new Error('🛑🤠 Not authorized 😡❌')
-    // }
+    const repair = await Repair.findById(req.params.repairId)
+    if (repair.owner.equals(req.user.profile)) {
+      await Repair.findByIdAndDelete(req.params.repairId)
+      const profile = await Profile.findById(req.user.profile)
+      profile.repairs.remove({ _id: req.params.repairId })
+      await profile.save()
+      res.json(repair)
+    } else {
+      throw new Error('🛑🤠 Not authorized 😡❌')
+    }
   } catch (err) {
     console.log(`🚨`, err)
     res.status(500).json(`🚨`, err)
